Add doxygen API documentation.
Add gh-pages to forked repo:
$ git checkout upstream/gh-pages -b gh-pages
$ git push -u origin gh-pages

Generate new doxygen docs:
$ git checkout master
$ ant wsclean
$ ant docs
$ git checkout gh-pages
$ rm -rf docs/doxygen
$ cp -r docs/generated/html/ docs/doxygen
$ git add docs/doxygen/

I think that the gh-pages docs needs to live in a different directory than
the master branch version, otherwise after you generate a new version of the
docs from the master branch, switching to the gh-pages branch will cause the
workspace to be reset back to what it looks like in that branch. Having separate
directories should get around this.