Final library reorg for now.
Single file lang.m contains core language support, is intended to
be implicitly imported into every script/module. Once we support
import circularity, this can be broken up if desired; until then
the dependencies make factoring impractical.

Common auxilliary stuff is collected in std.m, which is currently
loaded on shell startup. There's not much, mostly because non-toy
implementations of most things are blocked on variants and
interfaces.

It remains to finish grooming the code--all library functions
should be type annotated, and in some cases the docs could
stand to be clarified, or are missing entirely.